Wait for carrier on the bonding (bond0) interface.
Usually the init process is faster than the availability of the
established bonding network link. This results in a mount failure of the
GlusterFS volume, respective the netmount init script. The network
configuration now waits until the carrier on bond0 is up (with a timeout
of 15 seconds). This fixes bug #171